pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rmbg"
version = "0.1.1"
authors = [
    { name = "ShinChven", email = "shinchven@gmail.com" },
]
description = "A CLI tool to remove background from people's images"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "rembg",
    "Pillow",
]

[project.scripts]
rmbg = "rmbg.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["rmbg*"]